+++ packages/eas-cli/src/commands/update/view.ts
@@ -1,16 +1,23 @@
-import { Args } from '@oclif/core';
+import { Args, Flags } from '@oclif/core';
 import chalk from 'chalk';
 
 import EasCommand from '../../commandUtils/EasCommand';
 import { EasJsonOnlyFlag } from '../../commandUtils/flags';
+import { UpdateInsightsQuery } from '../../graphql/queries/UpdateInsightsQuery';
 import { UpdateQuery } from '../../graphql/queries/UpdateQuery';
 import Log from '../../log';
+import {
+  buildUpdateInsightsJson,
+  buildUpdateInsightsTable,
+  toUpdateInsightsSummary,
+} from '../../update/insights/formatInsights';
 import {
   formatUpdateGroup,
   getUpdateGroupDescriptions,
   getUpdateJsonInfosForUpdates,
 } from '../../update/utils';
 import { enableJsonOutput, printJsonOnlyOutput } from '../../utils/json';
+import { resolveInsightsTimeRange } from './insights';
 
 export default class UpdateView extends EasCommand {
   static override description = 'update group details';
@@ -23,6 +30,24 @@ export default class UpdateView extends EasCommand {
   };
 
   static override flags = {
+    insights: Flags.boolean({
+      description:
+        'Also show insights (launches, crash rate, unique users, payload size) for the update group.',
+      default: false,
+    }),
+    days: Flags.integer({
+      description: 'Show insights from the last N days (default 7). Only used with --insights.',
+      min: 1,
+      exclusive: ['start', 'end'],
+    }),
+    start: Flags.string({
+      description: 'Start of insights time range (ISO date). Only used with --insights.',
+      exclusive: ['days'],
+    }),
+    end: Flags.string({
+      description: 'End of insights time range (ISO date). Only used with --insights.',
+      exclusive: ['days'],
+    }),
     ...EasJsonOnlyFlag,
   };
 
@@ -33,7 +58,7 @@ export default class UpdateView extends EasCommand {
   async runAsync(): Promise<void> {
     const {
       args: { groupId },
-      flags: { json: jsonFlag },
+      flags: { json: jsonFlag, insights: insightsFlag, days, start, end },
     } = await this.parse(UpdateView);
 
     const {
@@ -46,14 +71,41 @@ export default class UpdateView extends EasCommand {
 
     const updatesByGroup = await UpdateQuery.viewUpdateGroupAsync(graphqlClient, { groupId });
 
+    const insightsSummary = insightsFlag
+      ? await (async () => {
+          const { daysBack, startTime, endTime } = resolveInsightsTimeRange({ days, start, end });
+          const updatesWithInsights = await UpdateInsightsQuery.viewUpdateGroupInsightsAsync(
+            graphqlClient,
+            { groupId, startTime, endTime }
+          );
+          return toUpdateInsightsSummary(groupId, updatesWithInsights, {
+            startTime,
+            endTime,
+            daysBack,
+          });
+        })()
+      : null;
+
     if (jsonFlag) {
-      printJsonOnlyOutput(getUpdateJsonInfosForUpdates(updatesByGroup));
+      if (insightsSummary) {
+        printJsonOnlyOutput({
+          updates: getUpdateJsonInfosForUpdates(updatesByGroup),
+          insights: buildUpdateInsightsJson(insightsSummary),
+        });
+      } else {
+        printJsonOnlyOutput(getUpdateJsonInfosForUpdates(updatesByGroup));
+      }
     } else {
       const [updateGroupDescription] = getUpdateGroupDescriptions([updatesByGroup]);
 
       Log.log(chalk.bold('Update group:'));
 
       Log.log(formatUpdateGroup(updateGroupDescription));
+
+      if (insightsSummary) {
+        Log.addNewLineIfNone();
+        Log.log(buildUpdateInsightsTable(insightsSummary));
+      }
     }
   }
 }